fix: responsive header
